requirements.txt
torch>=2.2.0
torchvision>=0.17.0
datasets>=2.19.0
nltk>=3.8.1
loralib>=0.1.2
PyYAML>=6.0.1
numpy>=1.26.0
tqdm>=4.66.0
rich>=13.7.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# 测试相关
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
